+++ changelog.md
@@ -24,11 +24,14 @@ Work in progress...
 
 ### Added
 - Add input array shape ensurance functions and start to use in sift & cycle submodules  [!26](https://gitlab.com/emd-dev/emd/-/merge_requests/26)
-
-### Fixed
+- Add more stopping criteria to sift module [!27](https://gitlab.com/emd-dev/emd/-/merge_requests/26)
+  - Rilling et al and fixed iterations IMF stopping criteria
+  - Energy threshold sift stopping criterion
 
 ### Changed
 - Refactor some options extrema detection functions [!29](https://gitlab.com/emd-dev/emd/-/merge_requests/29)
+- Sift throws an error if an IMF doesn't converge after a specified maximum number of iterations.
+- Refactor mask generation in mask sift. Now specifies N masks of different phases and has options for parallel processing.
 
 ---